+++ changelog.md
@@ -1,5 +1,16 @@
 # Manual change log
 
+## Latest
+
+
+## 2025.3.28.0
+
+- fix: base/extensions.py - improve idempotency #441
+- fix: base/container_ext/repo.py - user and secret are not required (although documentation states they are)
+- feature: base/admin.py - improve idempotency, support for new parameters in IVIA 11
+- feature: base/ssl_certificates/personal_certificate.py - support label parameter in IVIA 11
+- PEP 639: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license
+
 ## 2025.3.14.0
 
 - build: add pylint configuration